[Jobs] [CI] Deflake `test_sdk.py` by setting test env vars for head node, not just workers (#29806)
Why are these changes needed?
The environment variables for a test were supposed to be set for all workers in a cluster. However, the head node of the cluster was started by a pytest fixture before the call to monkeypatch.setenv that sets the environment variables.
Thus, the environment variables were only being set in the worker nodes, since they were started after monkeypatch.setenv.

This PR adds a fixture to set environment variables before starting the cluster head and uses that fixture in the tests.

It's likely that this will fix the flakiness, but this should be fixed regardless.

Related issue number
May address #29006

Signed-off-by: Weichen Xu <weichen.xu@databricks.com>