Adding logs to help debug Windows wheels flaky failure (#29878)
Windows Wheels build is flaky and it looks like it has something to do with our usage of refreshenv. From my reading, refreshenv basically exports the powershell environment into bash. It has some special case handling for environment variables which have lparens in their name (I guess this is a thing in Windows?). But from looking at recent flaky failures, I think it is missing some special case.

I thought it might fail for cases where the commit message has a single quote in it, but I ran the script locally (on powershell on mac) and it succeeded. Before spending more time on an investigation, I think it'll be easier to just get the exact case that's failing.

Related issue: #29717

Signed-off-by: Weichen Xu <weichen.xu@databricks.com>