[core] Fix bugs in ray syncer about warning and graceful destructing. (#29866)
There are two issues fixed in this PR:

In the current message flow, warning is pushed from raylet to gcs and then to autoscaler. This flow is really bad since it couple a lot of components for a simple feature (raylet/gcs/autoscaler) and ideally it should just be put in autoscaler.

Ray syncer failed in this feature because the state is not carried to autoscaler. To fix this, we'd better fix the flow, but given the complexity, we tried to fix it in syncer.

The new state is carried and pushed to autoscaler as previous.

The other one is about graceful exits. When the structure got shutdown, we need to make sure all init requests exits graceful to avoid crash.

Signed-off-by: Weichen Xu <weichen.xu@databricks.com>