[air/wandb] Deprecate Wandb mixin, move to `setup_wandb()` function (#29828)
The wandb trainable mixin is clunky to use and can't be used with training loops of Ray AIR trainers. Instead we introduce a utility method `setup_wandb()` which initializes a wandb session with (overwritable) default values. This method achieves maximum flexibility for users to interact with the wandb API as they are used to.

Additionally, we return a mock-API object for non rank-zero workers so that logging is not duplicated in distributed settings. This can also be disabled.

Further, this PR moves the wandb integration into `ray.air.integrations.wandb` and restructures the code so that the callback imports from this location.

As a note, this PR introduces `session.get_experiment_name()` as a way to retrieve the experiment name of the current run - this is needed to setup the default group name and seems to be a non-intrusive addition to the session API. But open to suggestions here.

Signed-off-by: Kai Fricke <kai@anyscale.com>
Signed-off-by: Weichen Xu <weichen.xu@databricks.com>