[Core] set worker non recyclable if gpu is envolved by default (#30061)
Current Ray behavior is to reuse Python interpreters across multiple Ray tasks to reduce latency. However, this comes with a lot of baggage for GPU tasks. Specifically, libraries like CUDA (and by extension, PyTorch and TensorFlow) tend not to be very mindful about clearing up GPU memory when references to objects holding GPU memory are removed by the Python garbage collector. For example: #29606

The workaround is to always set max_calls=1 on your Ray task to ensure the Python interpreter is shut down when the task completes. This frees the GPU memory and allows subsequent tasks to allocate additional memory.

We should by-default disable worker reuse for GPU tasks.
See #29624 for more context.

Signed-off-by: Stephanie Wang <swang@cs.berkeley.edu>

Signed-off-by: Weichen Xu <weichen.xu@databricks.com>