[Datasets] Only enable blocks bundling when batch_size is set (#29971)
Before this PR, we always enable blocks bundling in map_batches, to bundle small blocks together for the given batch_size. This is good for batch prediction on GPU, but not good for CPU preprocessing with default batch size (4096, which is too large). So here we decide to disable blocks bundling by default, and only enable blocks bundling when user specifies batch_size.

Signed-off-by: Weichen Xu <weichen.xu@databricks.com>