[Agent] Fix the agent logs streamed to drivers. (#30177)
Signed-off-by: SangBin Cho <rkooo567@gmail.com>

Right now, when ray is initialized, it starts its own logger (getLogger("ray")) with a handler (streamHandler) that's not controllable by users. When the ray.init is called within an agent, all the children loggers start sending logs to this "ray" logger, which prints all logs to stderr. Since the agent is a child of raylet, all stderr are redirected to the parent stderr, which is raylet.err. All the logs in raylet.err are printed to the driver by the log monitor.

This PR fixes the issue by not configuring the default ray logger when ray.init is called.

Signed-off-by: Weichen Xu <weichen.xu@databricks.com>