[Jobs] Add DELETE endpoint (#30056)
Adds a DELETE endpoint to the REST API, SDK and CLI for Ray Jobs.

Before this, there was no way to delete the job info from the cluster, so the memory could potentially grow without bound.

Deleting a job immediately deletes all its associated info (status, metadata) from memory. We only support deleting jobs that are in a terminal state.

We don't delete the log files of a deleted job.

Signed-off-by: Weichen Xu <weichen.xu@databricks.com>