[release][no_ci] Remove torch tune big gpu weekly test. (#30309)
The torch tune nightly GPU test is running 4 trials with 4 distributed workers each.
The corresponding weekly test (that the PR is removing) is running 4 trials with 8 distributed workers each.
To compare train with tune in a meaningful way, a "strict-pack" placement group strategy is needed. Hence, we need instances with 8 GPUs in order for the 8 workers to be strictly packed together.

Given this additional weekly test is not providing much value on top of the nightly test, proposing to remove it.

Signed-off-by: xwjiang2010 <xwjiang2010@gmail.com>
Signed-off-by: Weichen Xu <weichen.xu@databricks.com>