[Doc] [Jobs] Define driver and unify job definitions (#29608)
This is a minimal step towards unifying the definition of "Job" in the docs.

Right now there are two types of "Jobs": "Job Submission" jobs, and ordinary Ray drivers (which may or may not use Ray client.)

This PR updates the doc to emphasize that both of these are considered Jobs. To do this, we need to add a definition for driver, which this PR introduces.

Signed-off-by: Weichen Xu <weichen.xu@databricks.com>