[Tune] Make `Tuner.restore` work with relative experiment paths (#30363)
Relative paths can be set in RunConfig(local_dir="relative_path"), and the Experiment object that is created with this local_dir handles it properly by first expanding ~, then converting to an absolute path if needed.

The path passed into Tuner.restore(path) does not convert to an absolute path in the same way, which results in an error due to a mismatch between the Experiment local_dir (which is now an absolute path), and the relative experiment path that is supplied.

Signed-off-by: Justin Yu <justinvyu@berkeley.edu>
Signed-off-by: Weichen Xu <weichen.xu@databricks.com>