[dashboard] fix cpu_percent in reporter_agent (#29850)
In this pr, I revert the self._workers before #25418 (comment)_. The current one breaks the cpu_percent for all actors/tasks in the reporter agent.

Update: I believe the original intersection_update may be not suitable for the scenario due to its non-determinism. A naive implementation of intersection_update is included in this pr instead.

Co-authored-by: SangBin Cho <rkooo567@gmail.com>
Signed-off-by: 叶志晟 <yzs981130@126.com>
Signed-off-by: Weichen Xu <weichen.xu@databricks.com>