KubeRay node provider refactor (#30281)
Implements KubeRay node provider as a "BatchingNodeProvider".
Builds on #29933.

Summary of design
An autoscaler update now works like this:

list pod data from k8s
check if it's safe to proceed with update. Abort the update if not.
do some internal calculation to determine desired scale
submit a single patch to the RayCluster CR if a scale change is required
Everything is single-threaded and there are O(1) K8s API calls per autoscaler update.

Signed-off-by: Dmitri Gekhtman <dmitri.m.gekhtman@gmail.com>
Signed-off-by: Weichen Xu <weichen.xu@databricks.com>