[core] Set OMP_NUM_THREADS to `num_cpus` required by task/actors by default (#30496)
Ray currently sets OMP_NUM_THREADS=1 when the environ variable is not set.
This PR:

Sets OMP_NUM_THREADS to the number of cpus assigned to the worker that runs a task before running, and reset it after running.
If num_cpus is a fractional smaller than 1, it will set OMP_NUM_THREADS to 1.
Doesn't override OMP_NUM_THREADS if it's already being specified in runtime env or through os.environ.
Signed-off-by: Ricky Xu  <xuchen727@hotmail.com>
Co-authored-by: Eric Liang <ekhliang@gmail.com>
Co-authored-by: Simon Mo <simon.mo@hey.com>
Signed-off-by: Weichen Xu <weichen.xu@databricks.com>