[Test] Fix agent killed before raylet is terminated when `ray stop` is used (fix test_cli.py) (#30329)
Right now, when ray stop is called, it sends sigterm to all components.

It seems like after the recent dashboard optimization, the agent shutdown time became faster, and it shut down even faster than the raylet when sigterm is received.

This broke one of tests in test_cli.py which test if ray start --head --block is working as expected without unexpected component crashes (in this case, since the agent is killed earlier than raylet, raylet is considered unexpectedly dead).

I fix the issue by gracefully handling agent termination via SIGTERM. Basically, the agent will exit with exit code 0 (normal exit) upon sigterm. If the exit code is 0, raylet performs the graceful shutdown rather than QuickExit (raylet QuickExits when the agent is killed now). I am open to using other exit codes.

Note that the ideal solution is also to make raylet kills the agent when it gets the sigterm. I tried this first, but it is not very trivial, so I think we should do this as a follow up.

I also tried not terminating agent when ray stop is used (and relied on its fate sharing feature), but it fails lots of tests. I think it is because many tests rely on agent being killed immediately. Fixing this is non-trivial.

It also does a couple more things

The test current hangs in this scenario. I made it fail fast instead of hanging.
I found Drainself has a strong check when it is called more than once. But it seems like there's the case it is called more than once. So I made the method idempotent.

Signed-off-by: Weichen Xu <weichen.xu@databricks.com>