[core] add exponential backoff to task retry during oom (#30533)
Because the OOM killer kills task based on a pre-defined order, which is done to avoid live lock, it is possible that the same task gets re-submitted and killed, which may fail the task. Secondly, adding some form of retry delay will help produce back pressure and give time for longer running task to complete before it exhausts the retries.

Note this delay is only applied when the task fails due to OOM, and we maintain a separate delay for non-OOM retry that we may want to clean up (it is still used by a bunch of tests)
Signed-off-by: Clarence Ng <clarence.wyng@gmail.com>
Signed-off-by: Weichen Xu <weichen.xu@databricks.com>