[AIR] Ensure GBDT PGFs match passed ScalingConfig (#30470)
This PR fixes an issue with the special case from `GBDTTrainer`, where the `ScalingConfig` would be converted to `xgboost_ray.RayParams` first before being converted to a PGF for purposes of determining trial resources. This however results in dropping of strategy, args and kwargs from `ScalingConfig`, as `RayParams` doesn't support them. This PR introduces backwards compatible logic to deal with this issue. The changes introduced here will be upstreamed to `xgboost_ray`.

Signed-off-by: Antoni Baum <antoni.baum@protonmail.com>
Signed-off-by: Weichen Xu <weichen.xu@databricks.com>