[Train] Mention threading in `TensorflowTrainer` doc (#30543)
Even though we allow users to set the number of CPUs per worker, we do not set Tensorflow threading configurations automatically. This PR mentions this explicitly in the docstring.

We may consider setting this automatically in the future, but the main concern is that there are multiple configuration options interacting with another, so just setting them all to the number of CPUs assigned to a worker may not give the best performance. Therefore, we leave it up to the user for now.

Signed-off-by: Antoni Baum <antoni.baum@protonmail.com>
Signed-off-by: Weichen Xu <weichen.xu@databricks.com>