[doc] Add a session in ray core doc for tips to run large ray cluster. (#30599)
To run a large ray cluster, some parameters have to be tuned and also some OS configs have to be set. It requires a lot of experience for the users to figure out everything.
This PR add a session for this to help the user setup their large scale ray cluster.

Co-authored-by: Eric Liang <ekhliang@gmail.com>
Signed-off-by: Weichen Xu <weichen.xu@databricks.com>