[AIR] Change `ScalingConfig` to be optional for `DataParallelTrainer`s if already in Tuner `param_space` (#30715)
`DataParallelTrainer` currently throws an error when trying to initialize it without specifying `num_workers` in the scaling config. However, it's possible that the scaling config will be specified later in the Tuner `param_space`, in which case it may completely overwrite the original scaling config that's currently required. This PR moves this scaling config validation to happen after `trainer.fit`/`tuner.fit` is called. In the `tuner.fit` case, this would happen after the config has been merged with the one from the Tuner param space.

Signed-off-by: Justin Yu <justinvyu@berkeley.edu>
Signed-off-by: Weichen Xu <weichen.xu@databricks.com>