[core][state] Task events backend - worker task event buffer implementation [1/n] (#30867)
In this PR:

Added a TaskEventBuffer class which serves as an abstraction to store task events, and push those events to GCS in batches periodically.
Each CoreWorker will own one single TaskEventBuffer, and events (both task status change events and profiling events) will be added to a local in-memory buffer.
The TaskEventBuffer also owns its own GCS client and io thread, which is independent from the main io_contexts.

Signed-off-by: Weichen Xu <weichen.xu@databricks.com>