[CI] Update horovod, install after `ml_docker` (#30919)
Horovod has to be installed last (after torch and other libraries), but in our GPU tests we were installing it before ML Docker requirements, which contained DL libraries. This PR fixes the order and also updates the Horovod commit to the latest one - the old commit was from January 2022.

Not sure if it's worth centralizing the horovod commit in one place.

Signed-off-by: Antoni Baum <antoni.baum@protonmail.com>
Signed-off-by: Weichen Xu <weichen.xu@databricks.com>