Add doc about resource limits in docker (#30931)
The Kuberay quickstart does not work out of the box on a recent intel Macbook with default settings due to pods being stuck in PENDING because of insufficient memory. This is a quick PR to help unblock users who run into the error, and to prevent them from running into the error.

Tell the user to adjust the docker resource limits
Show the user how to debug pod stuck in "PENDING"

Signed-off-by: Archit Kulkarni <architkulkarni@users.noreply.github.com>
Signed-off-by: Weichen Xu <weichen.xu@databricks.com>