[core][scalability] Reduce unnecessary GCS get calls (16->8) in when starting a worker (#30166)
When a lot of workers started, it's going to add a lot of pressure to the GCS. Right now, we have 16 gcs internal kv get calls just for some metadata.

This PR optimize this and reduce it to 8 calls by passing the constants through the raylet directly.

Tests are added to enforce this and make sure no regression in the future.

Signed-off-by: Weichen Xu <weichen.xu@databricks.com>