Add top navigation to documentation (#30225)
This adds a top navigation to our documentation that will be unified with the top navigation of the ray.io website going forward so we can have a unified experience between ray.io website and documentation.

Before this is merged, @sijieamoy graciously offered to improve the styling of the top bar to fit better with the overall design. Currently the design is a placeholder with minimal work.

Co-authored-by: Sijie <sijie@anyscale.com>
Signed-off-by: Weichen Xu <weichen.xu@databricks.com>