[all_tests][ci] Use .zshenv to avoid conda setting up breaks the pipeline (#30999)
We have seen mac wheel build failures like when setting up the environment.

The error was raised when doing `source ~/.zshrc` which contains
- an export PATH to include bazel that `ci.sh init` puts in
- default conda activation code comes with the sandbox.

We don't use condas for the mac builds and tests I believe, and this PR moves the PATH export to `.zshenv` for environment.

(This also reverts the PR: https://github.com/ray-project/ray/pull/30911)

Signed-off-by: rickyyx <rickyx@anyscale.com>
Signed-off-by: Weichen Xu <weichen.xu@databricks.com>